+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
 	"programs/sp1/mock-consumer",
 	"programs/sp1/mock-consumer/program",
 	"programs/sp1/intensity-test/program",
+	"benches/vm_comparison",
 ]
 
 # Explicitly set the resolver to version 2, which is the default for packages with edition >= 2021
