+++ benches/README.md
@@ -1,0 +1,7 @@
+# VM Comparison Benchmarks
+
+This benchmark compares execution performance between RISC0 and SP1 zero-knowledge virtual machines.
+
+Currently, they both run a intensity test program with a various number of hash rounds. To run the benchmark, use `cargo run --bin vm-comparison`. 
+
+
