+++ benches/vm_comparison/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "vm-comparison"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+
+[lints]
+workspace = true
+
+[[bin]]
+name = "vm-comparison"
+path = "src/main.rs"
+
+[dependencies]
+risc0-zkvm = { workspace = true }
+sp1-zkvm = { workspace = true }
+alloy = { workspace = true }
+borsh = { workspace = true }
+sha2 = { version = "0.10", default-features = false }
+zkvm = { workspace = true }
